Improved handling of in-tweet URLs
Previously they were full of twitter-gunk, with nasty spaces and ellipses in the text content, and t.co hrefs and so on.
Now, they’re a clean link element with the contents of the href attribute as the text content.

Also added a test case specifically covering this, using a tweet from KartikPrabhu who spotted this problem.